requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.3.0
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
requests>=2.31.0
Flask>=2.3.0
FastAPI>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.5
